Reorganise files and classes
