+++ src/tools/behaviors/wheelBehavior.js
@@ -2,7 +2,7 @@ import {
   getMousePoint
 } from '../../gui/generic.js';
 import {logger} from '../../utils/logger.js';
-import {getActiveOrDrawRefViewLayer} from './panDragBehavior.js';
+import {getActiveOrDrawRefViewLayer} from './utils.js';
 
 // doc imports
 /* eslint-disable no-unused-vars */
